Simplify the format of animations as received by WebAssembly Studio
Currently animations are encoded to JSON by WebAssembly Studio and then immediately decoded into ImageData objects by the website. This patch changes the website to expect a single ArrayBuffer containing the entire animation, which is what WebAssembly Studio uses internally.

On the WebAssembly Studio side the new format can be sent with a simple template change.

The patch also removes the rotation of incoming frames, as we can adapt the templates to not need that. In fact that's much better because otherwise the mental model for how to map pixels to the Arch is more difficult.