Rotate frame data received from WebAssembly Studio by 180 degrees
The mapping of pixels to the Arch is highly unintuitive, and we want to avoid exposing this to users. The mapping is much easier to reason about if the pixels are rotated by 180 degrees, i.e. if the first pixel becomes the last, and vice versa.

The previewer has the inverse mapping, so to keep that the way it should be, we just rotate the 3D model by 180 degrees.